Fix: failing tests after merge

+++ server/src/test/java/server/api/GameControllerTest.java
@@ -250,13 +250,13 @@ public class GameControllerTest {
         sut.addRestartUser(gameIndex, 1);
         sut.addRestartUser(gameIndex, 2);
         sut.restartGame(gameIndex, (int) NUMBER, 1);
-        assertEquals(List.of((long) 2), sut.getGameList().getBody().getGames().get(gameIndex).getRestartUserIds());
+        assertEquals(List.of((long) 2), sut.getGameList().getBody().getGames().get((long) gameIndex).getRestartUserIds());
     }
 
     @Test
     public void restartGameSetsNewQuestionsOnFirstRequest() {
         Integer gameIndex = sut.startGame((int) NUMBER).getBody();
-        Game game = sut.getGameList().getBody().getGames().get(gameIndex);
+        Game game = sut.getGameList().getBody().getGames().get((long) gameIndex);
         List<Question> oldQuestions = game.getQuestions();
         sut.addRestartUser(gameIndex, 1);
         sut.restartGame(gameIndex, (int) NUMBER, 1);
@@ -266,7 +266,7 @@ public class GameControllerTest {
     @Test
     public void restartGameDoesNotSetNewQuestionsOnSecondRequest() {
         Integer gameIndex = sut.startGame((int) NUMBER).getBody();
-        Game game = sut.getGameList().getBody().getGames().get(gameIndex);
+        Game game = sut.getGameList().getBody().getGames().get((long) gameIndex);
         sut.addRestartUser(gameIndex, 1);
         sut.addRestartUser(gameIndex, 2);
         sut.restartGame(gameIndex, (int) NUMBER, 1);
@@ -278,7 +278,7 @@ public class GameControllerTest {
     @Test
     public void restartGameSetsNewQuestionsOnSecondRestart() {
         Integer gameIndex = sut.startGame((int) NUMBER).getBody();
-        Game game = sut.getGameList().getBody().getGames().get(gameIndex);
+        Game game = sut.getGameList().getBody().getGames().get((long) gameIndex);
         sut.addRestartUser(gameIndex, 1);
         sut.restartGame(gameIndex, (int) NUMBER, 1);
         List<Question> oldQuestions = game.getQuestions();
@@ -290,7 +290,7 @@ public class GameControllerTest {
     @Test
     public void restartGameReturnsFirstQuestion() {
         Integer gameIndex = sut.startGame((int) NUMBER).getBody();
-        Game game = sut.getGameList().getBody().getGames().get(gameIndex);
+        Game game = sut.getGameList().getBody().getGames().get((long) gameIndex);
         sut.addRestartUser(gameIndex, 1);
         Question returned = sut.restartGame(gameIndex, (int) NUMBER, 1).getBody();
         Question expected = game.getQuestions().get(0);
